Split styles in UserRoleDetailsDialog to resolve circular dependency
