+++ packages/sdk-react/src/components/userRoles/UserRoleDetails/styles.ts
@@ -1,0 +1,24 @@
+import { styled, TableCell, tableCellClasses } from '@mui/material';
+
+export const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.grey[100],
+    whiteSpace: 'nowrap',
+  },
+  [`&.${tableCellClasses.head}:first-of-type`]: {
+    position: 'sticky',
+    left: 0,
+    zIndex: '9999',
+    minWidth: '180px',
+    borderRight: `1px solid ${theme.palette.grey[300]}`,
+  },
+  [`&.${tableCellClasses.body}:first-of-type`]: {
+    position: 'sticky',
+    left: 0,
+    zIndex: '9998',
+    fontWeight: '600',
+    minWidth: '180px',
+    backgroundColor: theme.palette.background.default,
+    borderRight: `1px solid ${theme.palette.grey[300]}`,
+  },
+}));
